ServiceControl TA mixed transport sample
