+++ .gitignore
@@ -104,3 +104,4 @@ src/scaffolding.config
 .idea/
 *.sln.iml
 PromotedSqlScripts/
+/samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/.audit-db
