--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_1/Adapter/Adapter.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
-    <OutputType>Exe</OutputType>
-    <LangVersion>7.1</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="6.*" />
-    <PackageReference Include="NServiceBus.Raw" Version="2.*" />
-    <PackageReference Include="NServiceBus.SqlServer" Version="3.*" />
-    <PackageReference Include="ServiceControl.TransportAdapter" Version="1.*" />
-  </ItemGroup>
-</Project>
