--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_1/MixedTransportAdapter.sln
@@ -1,26 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.12
-MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sales", "Sales\Sales.csproj", "{29E5281B-F354-433C-A102-8C824206F031}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shipping", "Shipping\Shipping.csproj", "{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adapter", "Adapter\Adapter.csproj", "{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{29E5281B-F354-433C-A102-8C824206F031}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29E5281B-F354-433C-A102-8C824206F031}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
