--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_1/Shipping/Shipping.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
-    <OutputType>Exe</OutputType>
-    <LangVersion>7.1</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="6.*" />
-    <PackageReference Include="NServiceBus.SqlServer" Version="3.*" />
-    <PackageReference Include="NServiceBus.Heartbeat" Version="2.*" />
-  </ItemGroup>
-</Project>
