--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_1/Shipping/ShippingMessage.cs
@@ -1,7 +1,0 @@
-using NServiceBus;
-
-class ShippingMessage :
-    IMessage
-{
-    public string Id { get; set; }
-}
