+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Endpoint/ChaosGenerator.cs
