+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Endpoint/Endpoint.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net461</TargetFramework>
     <OutputType>Exe</OutputType>
@@ -6,6 +6,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="NServiceBus" Version="7.*" />
-    <PackageReference Include="NServiceBus.Transport.Msmq" Version="1.*" />
+    <PackageReference Include="NServiceBus.SqlServer" Version="5.*" />
+    <PackageReference Include="NServiceBus.Heartbeat" Version="3.*" />
   </ItemGroup>
 </Project>
