+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Endpoint/MyMessage.cs
@@ -1,6 +1,6 @@
 ﻿using NServiceBus;
 
-class SalesMessage :
+class MyMessage :
     IMessage
 {
     public string Id { get; set; }
