+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/MixedTransportAdapter.sln
@@ -1,26 +1,36 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.12
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29215.179
 MinimumVisualStudioVersion = 15.0.26730.12
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sales", "Sales\Sales.csproj", "{29E5281B-F354-433C-A102-8C824206F031}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Adapter", "Adapter\Adapter.csproj", "{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shipping", "Shipping\Shipping.csproj", "{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlatformLauncher", "PlatformLauncher\PlatformLauncher.csproj", "{8EDE58D6-68A4-4590-A383-6DFB7B482538}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adapter", "Adapter\Adapter.csproj", "{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Endpoint", "Endpoint\Endpoint.csproj", "{44CC7A25-B75E-4EAE-8DE9-DDC127270E4F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{29E5281B-F354-433C-A102-8C824206F031}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{29E5281B-F354-433C-A102-8C824206F031}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F8E8FC3E-0EDA-4B77-A2DE-A3C35DF616AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2C1EA06-4CC2-4899-B152-D8C4689A61DB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8EDE58D6-68A4-4590-A383-6DFB7B482538}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8EDE58D6-68A4-4590-A383-6DFB7B482538}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8EDE58D6-68A4-4590-A383-6DFB7B482538}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8EDE58D6-68A4-4590-A383-6DFB7B482538}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44CC7A25-B75E-4EAE-8DE9-DDC127270E4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44CC7A25-B75E-4EAE-8DE9-DDC127270E4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44CC7A25-B75E-4EAE-8DE9-DDC127270E4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44CC7A25-B75E-4EAE-8DE9-DDC127270E4F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {064B3DE1-30E8-4668-8CD4-F238E85C14FB}
 	EndGlobalSection
 EndGlobal
