+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/PlatformLauncher/PlatformLauncher.Core.csproj
@@ -1,11 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
     <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
     <LangVersion>7.1</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="6.*" />
-    <PackageReference Include="NServiceBus.SqlServer" Version="3.*" />
+    <PackageReference Include="Particular.PlatformSample" Version="1.*" />
   </ItemGroup>
 </Project>
