+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/PlatformLauncher/PlatformLauncher.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net461</TargetFramework>
+    <LangVersion>7.1</LangVersion>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Particular.PlatformSample" Version="1.*" />
+  </ItemGroup>
+</Project>
