+++ samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/PlatformLauncher/Program.cs
@@ -1,0 +1,13 @@
+﻿namespace PlatformLauncher
+{
+    using System;
+
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.Title = "PlatformLauncher";
+            Particular.PlatformLauncher.Launch();
+        }
+    }
+}
