--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Sales/SalesMessage.cs
@@ -1,7 +1,0 @@
-﻿using NServiceBus;
-
-class SalesMessage :
-    IMessage
-{
-    public string Id { get; set; }
-}
