--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Shipping/ChaosGenerator.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-public class ChaosGenerator
-{
-    public bool IsFailing { get; set; }
-
-    public Task Invoke()
-    {
-        if (IsFailing)
-        {
-            throw new Exception("Database is down");
-        }
-        return Task.CompletedTask;
-    }
-}
