--- samples/servicecontrol/adapter-mixed-transports/SCTransportAdapter_2/Shipping/Shipping.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
-    <OutputType>Exe</OutputType>
-    <LangVersion>7.1</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="7.*" />
-    <PackageReference Include="NServiceBus.SqlServer" Version="4.*" />
-    <PackageReference Include="NServiceBus.Heartbeat" Version="3.*" />
-    <PackageReference Include="System.Data.SqlClient" Version="4.*" />
-  </ItemGroup>
-</Project>
